renamed files to names in figma, fixed the UI in u2 profile init (have to remove =None from parent to run from program, changed UI on u3, fixed UI on u4 home
