+++ ui/pyqt/p6_care_profile_settings.py
