+++ ui/pyqt/u6_Emotion_session_sad.py
