+++ ui/pyqt/u7_camera_working_session_dashboard.py
