+++ ui/pyqt/u8_Emotion_session_annoyed.py
